fix: move skill-versions config under src/ so it ships with the function
Same class of bug as the CHANGELOG.md fix: config/skill-versions.ts sat
outside src/, and every other file the live app imports lives under
src/**, which is the only tree Vercel's Node.js function deployment for
this project ships (see AGENTS.md's known deployment constraints). No
existing src/*.ts file imported anything from outside src/ before this
change, and this was the one exception introduced earlier in this
branch. Relocates it to src/skill-versions.ts with no change to its
content or the public contract it discloses.
